Move RpcModule inside channel.rs
